+++ packages/hedron-core/src/hedron_core/_html_meta.py
@@ -232,6 +232,7 @@ ALLOWED_ATTRS = (
             "tabindex",
             "charset",
             "content",
+            "property",
             "http-equiv",
             "lang",
             "dir",